+++ tests/hlsl2glsltest/hlsl2glsltest.cpp
@@ -389,6 +389,7 @@ static bool CheckGLSL (bool vertex, ETargetVersion version, const std::string& s
 		replace_string (newSrc, "#extension GL_EXT_shadow_samplers : require", "", 0);
 		replace_string (newSrc, "#extension GL_EXT_frag_depth : require", "", 0);
 		replace_string (newSrc, "#extension GL_EXT_shader_framebuffer_fetch : require", "", 0);
+		replace_string (newSrc, "#extension GL_EXT_draw_buffers : require", "", 0);
 					   
 		sourcePtr = newSrc.c_str();
 	}